Added types to the form array.
